Move `getCookie` to `utils` module and add `setCookie` there as well
For consistency.
